split miniQuiz for 가독성성
